model_module.py
My output files have 66 vertical levels and you had written 50 at line 266. Because of that it was unable  to do the comparison at 272 and 273 due to different shapes.
I do not know if the vertical levels vary depending the model output. If that is the case that value should be not hard coded, right?